+++ CHANGELOG.md
@@ -32,6 +32,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix the `argv[0]` value for the [glibc] [dynamic loader][ld.so(8)] that does
   not support the option `--argv0` before 2.33; the value in `argv[0]` is now
   stripped from the path to the chroot
+- Fix the use for the [dynamic loader][ld.so(8)] option `--preload` which is
+  supported since [glibc] 2.30
 
 ### Security